add PublicAllocator tests, clean up

+++ test/helpers/BaseTest.sol
@@ -29,7 +29,7 @@ import "../../lib/forge-std/src/console2.sol";
 uint256 constant BLOCK_TIME = 1;
 uint256 constant MIN_TEST_ASSETS = 1e8;
 uint256 constant MAX_TEST_ASSETS = 1e28;
-uint184 constant CAP = type(uint128).max;
+uint184 constant CAP = type(uint112).max;
 uint256 constant NB_MARKETS = ConstantsLib.MAX_QUEUE_LENGTH + 1;
 
 contract BaseTest is EVaultTestBase {
@@ -69,7 +69,7 @@ contract BaseTest is EVaultTestBase {
             factory.createProxy(address(0), true, abi.encodePacked(address(loanToken), address(oracle), unitOfAccount))
         );
         eVault.setHookConfig(address(0), 0);
-        
+
         idleVault = _toIERC4626(eVault);
         vm.label(address(idleVault), "IdleVault");
         perspective.perspectiveVerify(address(idleVault));
